Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/css/cssom-view/scrollWidthHeightWhenNotScrollable-expected.txt


PASS elemSimple.clientHeight is the height of the padding edge
PASS elemSimple.scrollHeight is its clientHeight
PASS elemSimple.clientWidth is the width of the padding edge
PASS elemSimple.scrollWidth is its clientWidth
PASS elemOverflow.clientHeight is the height of the padding edge
PASS elemOverflow.scrollHeight is the height of its scrolled contents (ignoring padding, since we overflowed)
PASS elemOverflow.clientWidth is the width of the padding edge
PASS elemOverflow.scrollHeight is the width of its scrolled contents (ignoring padding, since we overflowed)
PASS elemNestedOverflow.clientHeight is the height of the padding edge
PASS elemNestedOverflow.scrollHeight is the height of its scrolled contents (ignoring padding, since we overflowed)
PASS elemNestedOverflow.clientWidth is the height of the padding edge
PASS elemNestedOverflow.scrollWidth is the width of its scrolled contents (ignoring padding, since we overflowed)
